[Update] Rename ExchangedDocumentContext to TaxInvoiceContext
