+++ src/main/java/th/ac/kmitl/it/soa/group3/model/taxinvoicecontext/TaxInvoiceContextModel.java
@@ -1,4 +1,4 @@
-package th.ac.kmitl.it.soa.group3.model.exchangeddocumentcontext;
+package th.ac.kmitl.it.soa.group3.model.taxinvoicecontext;
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
@@ -6,8 +6,8 @@ import lombok.Builder;
 
 @JacksonXmlRootElement(localName = "rsm:ExchangedDocumentContext")
 @Builder(builderClassName = "builder")
-public class ExchangedDocumentContextModel {
+public class TaxInvoiceContextModel {
 
     @JacksonXmlProperty
-    public GuidelineSpecifiedDocumentContextParameterModel  guidelineSpecifiedDocumentContextParameterModel;
+    public TaxInvoiceContextSpecModel taxInvoiceContextSpecModel;
 }
