+++ src/main/java/th/ac/kmitl/it/soa/group3/model/taxinvoicecontext/TaxInvoiceContextSpecModel.java
@@ -1,4 +1,4 @@
-package th.ac.kmitl.it.soa.group3.model.exchangeddocumentcontext;
+package th.ac.kmitl.it.soa.group3.model.taxinvoicecontext;
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
@@ -6,7 +6,7 @@ import lombok.Builder;
 
 @JacksonXmlRootElement(localName = "ram:GuidelineSpecifiedDocumentContextParameter")
 @Builder(builderClassName = "builder")
-public class GuidelineSpecifiedDocumentContextParameterModel {
+public class TaxInvoiceContextSpecModel {
 
     @JacksonXmlProperty(localName = "ram:ID")
     public String id;
